# docs/openapi.yaml
openapi: 3.1.0
info:
  title: piper API
  version: 0.1.0
  description: |
    Project-owned resources are always exposed below `/api/projects/{project_id}`.
    Collection POST operations return `201 Created`, successful mutations with no
    response body return `204 No Content`, JSON field names use snake_case, and
    errors use the shape `{"error":"message"}`.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/projects:
    get:
      description: Projects visible to the caller (all projects for a system admin).
      responses:
        "200":
          description: Projects.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Project" }
    post:
      description: System admin only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, name]
              properties:
                id: { type: string, pattern: "^[a-z0-9-]+$" }
                name: { type: string }
                description: { type: string }
                owner_member_id: { type: string }
      responses:
        "201":
          description: Created project.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
...
  /api/projects/{project_id}/members:
    parameters:
      - $ref: "#/components/parameters/ProjectID"
    get:
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Project members.
          headers:
            X-Total-Count: { $ref: "#/components/headers/TotalCount" }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ProjectMember" }
    post:
      description: Project admin only. Identify the user by `username` (or `user_id`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                user_id: { type: string }
                username: { type: string }
                role: { $ref: "#/components/schemas/ProjectRole" }
      responses:
        "201":
          description: Added member.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectMember" }
...
  /api/projects/{project_id}/members/{user_id}:
    parameters:
      - $ref: "#/components/parameters/ProjectID"
      - name: user_id
        in: path
        required: true
        schema: { type: string }
    patch:
      description: Project admin only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { $ref: "#/components/schemas/ProjectRole" }
      responses:
        "200":
          description: Updated member.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectMember" }
    delete:
      description: Project admin only.
      responses:
        "204":
          description: Removed.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    ProjectID:
      name: project_id
      in: path
      required: true
      schema: { type: string }
...
    Limit:
      name: limit
      in: query
      description: Page size. When set, the response carries `X-Total-Count`.
      schema: { type: integer, minimum: 1 }
    Offset:
      name: offset
      in: query
      description: Rows to skip; only applies together with `limit`.
      schema: { type: integer, minimum: 0 }
  headers:
    TotalCount:
      description: Total rows matching the filter, ignoring limit/offset. Sent whenever `limit` is.
      schema: { type: integer }
  schemas:
...
    ProjectRole:
      type: string
      enum: [viewer, member, admin]
    ProjectMember:
      type: object
      properties:
        project_id: { type: string }
        user_id: { type: string }
        username: { type: string }
        role: { $ref: "#/components/schemas/ProjectRole" }
    Project:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        owner_member_id: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...