# memberclient_local.go
package piper

import (
	"bytes"
	"context"
	"crypto/sha256"
	"encoding/base64"
	"encoding/json"
	"errors"
	"fmt"
	"log/slog"
	"net/http"
	"net/http/httptest"
	"net/url"
	"strings"
	"time"

	"github.com/loykin/piper/internal/logstore"
	"github.com/loykin/piper/internal/memberclient"
	"github.com/loykin/piper/internal/projectclient"
	"github.com/loykin/piper/pkg/pipeline/run"
	"github.com/loykin/piper/pkg/project"
	"github.com/loykin/piper/pkg/security"
	"github.com/loykin/piper/pkg/statsstore"
)

// localMemberClient implements memberclient.Client for the single-install
// case: Home calls its one Local Member in-process (fed.md §13.11). It is a
// thin translation layer over *Piper's runs Manager (internal/runlifecycle) —
// pinning project context and unpacking wire DTOs, not duplicating run
// business logic (including SubmitRun's idempotency bookkeeping, which lives
// in Manager.SubmitRun, not here).
type localMemberClient struct {
	p              *Piper
	projectHandler http.Handler
}

type LocalMemberClient interface {
	memberclient.Client
	projectclient.Client
}

// NewLocalMemberClient wraps p to satisfy memberclient.Client in-process.
func NewLocalMemberClient(p *Piper) LocalMemberClient {
	client := &localMemberClient{p: p}
	client.projectHandler = p.newMemberProjectRouter()
	return client
}

// withProjectContext pins ctx's project.Context to exactly what Home
// resolved (ref/auth), rather than trusting whatever project.Context the
// incoming ctx already happens to carry — Member acts on what it was
// explicitly handed, matching fed.md §10.10's "Member validates the
// authorization context" invariant even though, for the in-process Local
// Member, Home and Member share the same request lifecycle.
func withProjectContext(ctx context.Context, auth memberclient.AuthContext, ref project.ProjectRef) context.Context {
	return project.WithContext(ctx, project.Context{ID: ref.ProjectID, Role: auth.Role})
}

func toRunSummary(r *run.Run) memberclient.RunSummary {
	version := r.VersionFromYAML()
	redacted := r.Redact()
	return memberclient.RunSummary{
		ID:              redacted.ID,
		ProjectID:       redacted.ProjectID,
		ScheduleID:      redacted.ScheduleID,
		Experiment:      redacted.Experiment,
		PipelineName:    redacted.PipelineName,
		PipelineVersion: version,
		Status:          redacted.Status,
		StartedAt:       redacted.StartedAt,
		EndedAt:         redacted.EndedAt,
		ScheduledAt:     redacted.ScheduledAt,
		PipelineYAML:    redacted.PipelineYAML,
		ParamsJSON:      redacted.ParamsJSON,
		CreatedBy:       redacted.CreatedBy,
	}
}

func toStepSummary(s *run.Step) memberclient.StepSummary {
	return memberclient.StepSummary{
		ProjectID: s.ProjectID,
		RunID:     s.RunID,
		StepName:  s.StepName,
		Status:    s.Status,
		StartedAt: s.StartedAt,
		EndedAt:   s.EndedAt,
		Error:     s.Error,
		Attempts:  s.Attempts,
	}
}

func toStepSummaries(steps []*run.Step) []memberclient.StepSummary {
	out := make([]memberclient.StepSummary, 0, len(steps))
	for _, s := range steps {
		out = append(out, toStepSummary(s))
	}
	return out
}

// SubmitRun unpacks the wire DTO, pins the project context, and delegates to
// the Manager's idempotency-aware SubmitRun — see internal/runlifecycle's
// doc comment for the SHA-256/Claim/replay flow this used to implement here.
func (l *localMemberClient) SubmitRun(ctx context.Context, auth memberclient.AuthContext, ref project.ProjectRef, req memberclient.SubmitRunRequest) (memberclient.SubmitRunResponse, error) {
	ctx = withProjectContext(ctx, auth, ref)
	runID, err := l.p.runs.SubmitRun(ctx, ref.ProjectID, req.IdempotencyKey, req.YAML, req.Params, req.Vars, req.Experiment)
	if err != nil {
		return memberclient.SubmitRunResponse{}, err
	}
	return memberclient.SubmitRunResponse{RunID: runID}, nil
}

func (l *localMemberClient) SubmitSweep(ctx context.Context, auth memberclient.AuthContext, ref project.ProjectRef, req memberclient.SubmitSweepRequest) (memberclient.SubmitSweepResponse, error) {
	ctx = withProjectContext(ctx, auth, ref)
	trials := make([]run.SweepTrial, 0, len(req.Runs))
	for _, t := range req.Runs {
		trials = append(trials, run.SweepTrial{Params: t.Params})
	}
	resp, err := l.p.runs.StartSweep(ctx, ref.ProjectID, run.SweepRequest{YAML: req.YAML, Experiment: req.Experiment, Runs: trials})
	if err != nil {
		return memberclient.SubmitSweepResponse{}, err
	}
	return memberclient.SubmitSweepResponse{Experiment: resp.Experiment, RunIDs: resp.RunIDs}, nil
}

func (l *localMemberClient) ListRuns(ctx context.Context, _ memberclient.AuthContext, ref project.ProjectRef, req memberclient.ListRunsRequest) (memberclient.ListRunsResponse, error) {
	filter := run.RunFilter{
		Experiment:   req.Experiment,
		PipelineName: req.PipelineName,
		ScheduleID:   req.ScheduleID,
		Status:       req.Status,
		MetricStep:   req.MetricStep,
		MetricKey:    req.MetricKey,
		MetricOrder:  req.MetricOrder,
		Limit:        req.Limit,
		Offset:       req.Offset,
	}
	runs, err := l.p.repos.Run.List(ctx, ref.ProjectID, filter)
	if err != nil {
		return memberclient.ListRunsResponse{}, err
	}
	resp := memberclient.ListRunsResponse{Runs: make([]memberclient.RunSummary, 0, len(runs))}
	for _, r := range runs {
		resp.Runs = append(resp.Runs, toRunSummary(r))
	}
	if req.Limit > 0 {
		total, err := l.p.repos.Run.Count(ctx, ref.ProjectID, filter)
		if err != nil {
			return memberclient.ListRunsResponse{}, err
		}
		resp.Total = total
	}
	if req.IncludeSteps {
		runIDs := make([]string, 0, len(runs))
		for _, r := range runs {
			runIDs = append(runIDs, r.ID)
		}
		stepsByRun, err := l.p.repos.Step.ListByRuns(ctx, ref.ProjectID, runIDs)
		if err != nil {
			return memberclient.ListRunsResponse{}, err
		}
		resp.Steps = make(map[string][]memberclient.StepSummary, len(stepsByRun))
		for runID, steps := range stepsByRun {
			resp.Steps[runID] = toStepSummaries(steps)
		}
	}
	return resp, nil
}

func (l *localMemberClient) ListExperiments(ctx context.Context, _ memberclient.AuthContext, ref project.ProjectRef, req memberclient.ListExperimentsRequest) (memberclient.ListExperimentsResponse, error) {
	rows, total, err := l.p.repos.Run.ListExperiments(ctx, ref.ProjectID, run.ExperimentFilter{Name: req.Name, PipelineName: req.PipelineName, Limit: req.Limit, Offset: req.Offset})
	if err != nil {
		return memberclient.ListExperimentsResponse{}, err
	}
	resp := memberclient.ListExperimentsResponse{Experiments: make([]memberclient.ExperimentSummary, 0, len(rows)), Total: total}
	for _, row := range rows {
		resp.Experiments = append(resp.Experiments, memberclient.ExperimentSummary{
			Name: row.Name, Runs: row.Runs, Success: row.Success, Failed: row.Failed, Running: row.Running, Latest: row.Latest,
		})
	}
	return resp, nil
}

func (l *localMemberClient) GetRun(ctx context.Context, _ memberclient.AuthContext, ref project.ProjectRef, runID string) (memberclient.RunDetail, error) {
	r, err := l.p.repos.Run.Get(ctx, ref.ProjectID, runID)
	if err != nil || r == nil {
		return memberclient.RunDetail{}, memberclient.ErrRunNotFound
	}
	steps, err := l.p.repos.Step.List(ctx, ref.ProjectID, runID)
	if err != nil {
		steps = nil // best-effort, matches the handler's prior slog.Warn-and-continue behavior
	}
	return memberclient.RunDetail{Run: toRunSummary(r), Steps: toStepSummaries(steps)}, nil
}

func (l *localMemberClient) CancelRun(ctx context.Context, auth memberclient.AuthContext, ref project.ProjectRef, runID string) error {
	ctx = withProjectContext(ctx, auth, ref)
	return l.p.CancelRun(ctx, runID)
}

func (l *localMemberClient) RerunRun(ctx context.Context, auth memberclient.AuthContext, ref project.ProjectRef, runID string, failedOnly bool) (string, error) {
	ctx = withProjectContext(ctx, auth, ref)
	return l.p.RerunRun(ctx, runID, failedOnly)
}

func (l *localMemberClient) ListSteps(ctx context.Context, _ memberclient.AuthContext, ref project.ProjectRef, runID string) ([]memberclient.StepSummary, error) {
	steps, err := l.p.repos.Step.List(ctx, ref.ProjectID, runID)
	if err != nil {
		return nil, err
	}
	return toStepSummaries(steps), nil
}

func (l *localMemberClient) RetryStep(ctx context.Context, auth memberclient.AuthContext, ref project.ProjectRef, runID, stepName string) (string, error) {
	ctx = withProjectContext(ctx, auth, ref)
	return l.p.RetryStep(ctx, runID, stepName)
}

func (l *localMemberClient) QueryLogs(ctx context.Context, _ memberclient.AuthContext, ref project.ProjectRef, req memberclient.QueryLogsRequest) (memberclient.QueryLogsResponse, error) {
	cursor := req.Cursor
	if cursor == "" && req.AfterID > 0 {
		cursor = statsstore.CursorFromID(req.AfterID)
	}
	page, err := l.p.stats.Logs.QueryLogs(ctx, statsstore.LogQuery{ProjectID: ref.ProjectID, RunID: req.RunID, StepName: req.StepName, Cursor: cursor, Since: req.Since, Until: req.Until, Search: req.Search, Limit: req.Limit})
	if err != nil {
		if errors.Is(err, statsstore.ErrInvalidCursor) {
			return memberclient.QueryLogsResponse{}, err
		}
		return memberclient.QueryLogsResponse{}, fmt.Errorf("%w: %v", statsstore.ErrBackendUnavailable, err)
	}
	lines := make([]*logstore.Line, len(page.Lines))
	for i := range page.Lines {
		lines[i] = &page.Lines[i]
	}
	return memberclient.QueryLogsResponse{Lines: lines, NextCursor: page.NextCursor}, nil
}

func (l *localMemberClient) StatsCapabilities(context.Context, memberclient.AuthContext, project.ProjectRef) (statsstore.Capabilities, error) {
	capabilities := l.p.stats.Capabilities
	health := l.p.stats.Health()
	capabilities.Healthy, capabilities.Degraded = health.Healthy, health.Degraded
	capabilities.PendingBytes, capabilities.LastError = health.PendingBytes, health.LastError
	return capabilities, nil
}

func (l *localMemberClient) PurgeProjectStats(ctx context.Context, _ memberclient.AuthContext, ref project.ProjectRef) error {
	return l.p.stats.PurgeProject(ctx, ref.ProjectID)
}

func (l *localMemberClient) QueryMetrics(ctx context.Context, _ memberclient.AuthContext, ref project.ProjectRef, req memberclient.QueryMetricsRequest) (memberclient.QueryMetricsResponse, error) {
	if l.p.stats.Metrics == nil {
		return memberclient.QueryMetricsResponse{}, nil
	}
	page, err := l.p.stats.Metrics.QueryMetrics(ctx, statsstore.MetricQuery{ProjectID: ref.ProjectID, RunID: req.RunID, StepName: req.StepName, Keys: req.Keys, Cursor: req.Cursor, Since: req.Since, Until: req.Until, Limit: req.Limit})
	if err != nil {
		if errors.Is(err, statsstore.ErrInvalidCursor) {
			return memberclient.QueryMetricsResponse{}, err
		}
		return memberclient.QueryMetricsResponse{}, fmt.Errorf("%w: %v", statsstore.ErrBackendUnavailable, err)
	}
	points := make([]*logstore.Metric, len(page.Points))
	for i := range page.Points {
		points[i] = &page.Points[i]
	}
	return memberclient.QueryMetricsResponse{Points: points, NextCursor: page.NextCursor}, nil
}

func (l *localMemberClient) ListArtifacts(ctx context.Context, auth memberclient.AuthContext, ref project.ProjectRef, runID string) ([]any, error) {
	ctx = withProjectContext(ctx, auth, ref)
	return (&piperArtifacts{p: l.p}).List(ctx, runID)
}

func (l *localMemberClient) ServeArtifact(ctx context.Context, auth memberclient.AuthContext, ref project.ProjectRef, w http.ResponseWriter, r *http.Request, runID, step, path string) {
	ctx = withProjectContext(ctx, auth, ref)
	r = r.WithContext(ctx)
	(&piperArtifacts{p: l.p}).ServeDownload(w, r, runID, step, path)
}

func (l *localMemberClient) DoProjectRequest(ctx context.Context, auth memberclient.AuthContext, ref project.ProjectRef, req projectclient.Request) (projectclient.Response, error) {
	if req.Path == "" || !strings.HasPrefix(req.Path, "/") || strings.Contains(req.Path, "..") {
		return projectclient.Response{}, fmt.Errorf("projectclient: invalid project-relative path %q", req.Path)
	}
	if err := l.ensureExecutionProject(ctx, ref.ProjectID); err != nil {
		return projectclient.Response{}, err
	}
	var mutation *projectclient.Mutation
	key := http.Header(req.Header).Get("Idempotency-Key")
	if key != "" && req.Method != http.MethodGet && req.Method != http.MethodHead && l.p.repos.ProjectMutation != nil {
		hashPayload, err := json.Marshal(struct {
			Method, Path, Query string
			Body                []byte
		}{req.Method, req.Path, req.RawQuery, req.Body})
		if err != nil {
			return projectclient.Response{}, err
		}
		sum := sha256.Sum256(hashPayload)
		mutation = &projectclient.Mutation{ProjectID: ref.ProjectID, Key: key, RequestHash: base64.RawURLEncoding.EncodeToString(sum[:]), CreatedAt: time.Now().UTC()}
		existing, claimed, err := l.p.repos.ProjectMutation.Claim(ctx, mutation)
		if err != nil {
			return projectclient.Response{}, err
		}
		if existing.RequestHash != mutation.RequestHash {
			return projectclient.Response{Status: http.StatusConflict, Body: []byte(`{"error":"idempotency key was already used for a different request"}`), Header: http.Header{"Content-Type": []string{"application/json"}}}, nil
		}
		if !claimed {
			if !existing.Completed {
				return projectclient.Response{Status: http.StatusConflict, Body: []byte(`{"error":"request is already in progress or its outcome is unknown; retry later or reconcile before issuing a new mutation"}`), Header: http.Header{"Content-Type": []string{"application/json"}}}, nil
			}
			var headers http.Header
			_ = json.Unmarshal(existing.HeaderJSON, &headers)
			return projectclient.Response{Status: existing.Status, Header: headers, Body: existing.Body}, nil
		}
	}
	target := "/projects/" + url.PathEscape(ref.ProjectID) + req.Path
	if req.RawQuery != "" {
		target += "?" + req.RawQuery
	}
	httpReq := httptest.NewRequestWithContext(ctx, req.Method, target, bytes.NewReader(req.Body))
	for name, values := range req.Header {
		for _, value := range values {
			httpReq.Header.Add(name, value)
		}
	}
	projectCtx := project.Context{ID: ref.ProjectID, OwnerMemberID: project.LocalMemberID, Role: auth.Role}
	httpCtx := project.WithContext(httpReq.Context(), projectCtx)
	if auth.ActorID != "" {
		httpCtx = security.WithIdentity(httpCtx, &security.Identity{ID: auth.ActorID})
	}
	httpReq = httpReq.WithContext(httpCtx)
	recorder := httptest.NewRecorder()
	l.projectHandler.ServeHTTP(recorder, httpReq)
	response := projectclient.Response{
		Status: recorder.Code,
		Header: recorder.Header().Clone(),
		Body:   append([]byte(nil), recorder.Body.Bytes()...),
	}
	if mutation != nil {
		mutation.Status = response.Status
		mutation.Body = append([]byte(nil), response.Body...)
		mutation.Completed = true
		mutation.HeaderJSON, _ = json.Marshal(response.Header)
		completeCtx, cancel := context.WithTimeout(context.WithoutCancel(ctx), 5*time.Second)
		defer cancel()
		if err := l.p.repos.ProjectMutation.Complete(completeCtx, mutation); err != nil {
			// The mutation itself already succeeded (response was built
			// above); only its idempotency record failed to persist. Return
			// the real result instead of hiding it behind this error - the
			// claim row stays reclaimable (see projectclient.StaleClaimWindow)
			// so a same-key retry can still complete it later.
			slog.Warn("project mutation completed but idempotency record failed to persist", "project_id", ref.ProjectID, "key", key, "err", err)
		}
	}
	return response, nil
}

func (l *localMemberClient) ServeProjectHTTP(ctx context.Context, auth memberclient.AuthContext, ref project.ProjectRef, w http.ResponseWriter, req *http.Request) error {
	if err := l.ensureExecutionProject(ctx, ref.ProjectID); err != nil {
		return err
	}
	clone := req.Clone(ctx)
	urlCopy := *clone.URL
	clone.URL = &urlCopy
	if strings.HasPrefix(clone.URL.Path, "/api/projects/") {
		clone.URL.Path = strings.TrimPrefix(clone.URL.Path, "/api")
	}
	projectCtx := project.Context{ID: ref.ProjectID, OwnerMemberID: project.LocalMemberID, Role: auth.Role}
	httpCtx := project.WithContext(clone.Context(), projectCtx)
	if auth.ActorID != "" {
		httpCtx = security.WithIdentity(httpCtx, &security.Identity{ID: auth.ActorID})
	}
	clone = clone.WithContext(httpCtx)
	l.projectHandler.ServeHTTP(w, clone)
	return nil
}

func (l *localMemberClient) ensureExecutionProject(ctx context.Context, projectID string) error {
	value, err := l.p.repos.Project.Get(ctx, projectID)
	if err != nil {
		return err
	}
	if value != nil {
		return nil
	}
	err = l.p.repos.Project.Create(ctx, &project.Project{
		ID: projectID, Name: projectID, OwnerMemberID: project.LocalMemberID,
	})
	if err == nil {
		return nil
	}
	value, getErr := l.p.repos.Project.Get(ctx, projectID)
	if getErr == nil && value != nil {
		return nil
	}
	return err
}

var _ memberclient.Client = (*localMemberClient)(nil)
var _ projectclient.Client = (*localMemberClient)(nil)
var _ projectclient.StreamClient = (*localMemberClient)(nil)